move property validation control API
